+++ src/include/stream.h
@@ -16,7 +16,7 @@ struct disk_stream
 
 struct disk_stream* create_disk_streamer(uint32_t disk_id);
 int disk_streamer_seek(struct disk_stream* stream, int pos);
-size_t disk_streamer_read(struct disk_stream* stream, void* out, int total);
+int disk_streamer_read(struct disk_stream* stream, void* out, int total);
 void disk_streamer_close(struct disk_stream* stream);
 
 #endif
